+++ test/test_full_reduced_density_matrix.py
@@ -4,14 +4,14 @@ from math import sqrt
 import numpy as np
 
 import pyqentangle
-import pyqentangle.tncompute
+import pyqentangle.core.tncompute
 import pytest
 
 
 def test_uneven_two_levels():
     tensor = np.array([[0., np.sqrt(np.reciprocal(3.))],
                        [-np.sqrt(2./3.)*1j, 0.]])
-    fulldenmat = pyqentangle.tncompute.bipartitepurestate_densitymatrix(tensor)
+    fulldenmat = pyqentangle.core.tncompute.bipartitepurestate_densitymatrix(tensor)
 
     assert fulldenmat[0, 1, 0, 1] == pytest.approx(np.reciprocal(3.))
     assert fulldenmat[1, 0, 1, 0] == pytest.approx(2/3)
